# Vectorise BitArray