--- a/src/System.Collections/src/System/Collections/BitArray.cs
+++ b/src/System.Collections/src/System/Collections/BitArray.cs
@@ -241,13 +241,8 @@ namespace System.Collections
         public void SetAll(bool value)
         {
             int fillValue = value ? -1 : 0;
-            int[] array = m_array;
-
-            for (int i = 0; i < array.Length; i++)
-            {
-                array[i] = fillValue;
-            }
-
+            int arrayLength = GetInt32ArrayLengthFromBitLength(Length);
+            m_array.AsSpan(0, arrayLength).Fill(fillValue);
             _version++;
         }
 
@@ -275,8 +270,13 @@ namespace System.Collections
             if (Length != value.Length || (uint)count > (uint)thisArray.Length || (uint)count > (uint)valueArray.Length)
                 throw new ArgumentException(SR.Arg_ArrayLengthsDiffer);
 
+            // Unroll loop for count less than Vector256 size.
             switch (count)
             {
+                case 7: thisArray[6] &= valueArray[6]; goto case 6;
+                case 6: thisArray[5] &= valueArray[5]; goto case 5;
+                case 5: thisArray[4] &= valueArray[4]; goto case 4;
+                case 4: thisArray[3] &= valueArray[3]; goto case 3;
                 case 3: thisArray[2] &= valueArray[2]; goto case 2;
                 case 2: thisArray[1] &= valueArray[1]; goto case 1;
                 case 1: thisArray[0] &= valueArray[0]; goto Done;
@@ -284,7 +284,20 @@ namespace System.Collections
             }
 
             int i = 0;
-            if (Sse2.IsSupported)
+            if (Avx2.IsSupported)
+            {
+                fixed (int* leftPtr = thisArray)
+                fixed (int* rightPtr = valueArray)
+                {
+                    for (; i < count - (Vector256<int>.Count - 1); i += Vector256<int>.Count)
+                    {
+                        Vector256<int> leftVec = Avx.LoadVector256(leftPtr + i);
+                        Vector256<int> rightVec = Avx.LoadVector256(rightPtr + i);
+                        Avx.Store(leftPtr + i, Avx2.And(leftVec, rightVec));
+                    }
+                }
+            }
+            else if (Sse2.IsSupported)
             {
                 fixed (int* leftPtr = thisArray)
                 fixed (int* rightPtr = valueArray)
@@ -330,8 +343,13 @@ namespace System.Collections
             if (Length != value.Length || (uint)count > (uint)thisArray.Length || (uint)count > (uint)valueArray.Length)
                 throw new ArgumentException(SR.Arg_ArrayLengthsDiffer);
 
+            // Unroll loop for count less than Vector256 size.
             switch (count)
             {
+                case 7: thisArray[6] |= valueArray[6]; goto case 6;
+                case 6: thisArray[5] |= valueArray[5]; goto case 5;
+                case 5: thisArray[4] |= valueArray[4]; goto case 4;
+                case 4: thisArray[3] |= valueArray[3]; goto case 3;
                 case 3: thisArray[2] |= valueArray[2]; goto case 2;
                 case 2: thisArray[1] |= valueArray[1]; goto case 1;
                 case 1: thisArray[0] |= valueArray[0]; goto Done;
@@ -339,7 +357,20 @@ namespace System.Collections
             }
 
             int i = 0;
-            if (Sse2.IsSupported)
+            if (Avx2.IsSupported)
+            {
+                fixed (int* leftPtr = thisArray)
+                fixed (int* rightPtr = valueArray)
+                {
+                    for (; i < count - (Vector256<int>.Count - 1); i += Vector256<int>.Count)
+                    {
+                        Vector256<int> leftVec = Avx.LoadVector256(leftPtr + i);
+                        Vector256<int> rightVec = Avx.LoadVector256(rightPtr + i);
+                        Avx.Store(leftPtr + i, Avx2.Or(leftVec, rightVec));
+                    }
+                }
+            }
+            else if (Sse2.IsSupported)
             {
                 fixed (int* leftPtr = thisArray)
                 fixed (int* rightPtr = valueArray)
@@ -385,8 +416,13 @@ namespace System.Collections
             if (Length != value.Length || (uint)count > (uint)thisArray.Length || (uint)count > (uint)valueArray.Length)
                 throw new ArgumentException(SR.Arg_ArrayLengthsDiffer);
 
+            // Unroll loop for count less than Vector256 size.
             switch (count)
             {
+                case 7: thisArray[6] ^= valueArray[6]; goto case 6;
+                case 6: thisArray[5] ^= valueArray[5]; goto case 5;
+                case 5: thisArray[4] ^= valueArray[4]; goto case 4;
+                case 4: thisArray[3] ^= valueArray[3]; goto case 3;
                 case 3: thisArray[2] ^= valueArray[2]; goto case 2;
                 case 2: thisArray[1] ^= valueArray[1]; goto case 1;
                 case 1: thisArray[0] ^= valueArray[0]; goto Done;
@@ -394,7 +430,20 @@ namespace System.Collections
             }
 
             int i = 0;
-            if (Sse2.IsSupported)
+            if (Avx2.IsSupported)
+            {
+                fixed (int* leftPtr = m_array)
+                fixed (int* rightPtr = value.m_array)
+                {
+                    for (; i < count - (Vector256<int>.Count - 1); i += Vector256<int>.Count)
+                    {
+                        Vector256<int> leftVec = Avx.LoadVector256(leftPtr + i);
+                        Vector256<int> rightVec = Avx.LoadVector256(rightPtr + i);
+                        Avx.Store(leftPtr + i, Avx2.Xor(leftVec, rightVec));
+                    }
+                }
+            }
+            else if (Sse2.IsSupported)
             {
                 fixed (int* leftPtr = thisArray)
                 fixed (int* rightPtr = valueArray)
@@ -421,15 +470,60 @@ namespace System.Collections
         ** off/false. Off/false bit values are turned on/true. The current instance
         ** is updated and returned.
         =========================================================================*/
-        public BitArray Not()
+        public unsafe BitArray Not()
         {
-            int[] array = m_array;
+            // This method uses unsafe code to manipulate data in the BitArray.  To avoid issues with
+            // buggy code concurrently mutating this instance in a way that could cause memory corruption,
+            // we snapshot the array then operate only on this snapshot.  We don't care about such code
+            // corrupting the BitArray data in a way that produces incorrect answers, since BitArray is not meant
+            // to be thread-safe; we only care about avoiding buffer overruns.
+            int[] thisArray = m_array;
 
-            for (int i = 0; i < array.Length; i++)
+            int count = GetInt32ArrayLengthFromBitLength(Length);
+
+            // Unroll loop for count less than Vector256 size.
+            switch (count)
             {
-                array[i] = ~array[i];
+                case 7: thisArray[6] = ~thisArray[6]; goto case 6;
+                case 6: thisArray[5] = ~thisArray[5]; goto case 5;
+                case 5: thisArray[4] = ~thisArray[4]; goto case 4;
+                case 4: thisArray[3] = ~thisArray[3]; goto case 3;
+                case 3: thisArray[2] = ~thisArray[2]; goto case 2;
+                case 2: thisArray[1] = ~thisArray[1]; goto case 1;
+                case 1: thisArray[0] = ~thisArray[0]; goto Done;
+                case 0: goto Done;
             }
 
+            int i = 0;
+            if (Avx2.IsSupported)
+            {
+                Vector256<int> ones = Vector256.Create(-1);
+                fixed (int* ptr = thisArray)
+                {
+                    for (; i < count - (Vector256<int>.Count - 1); i += Vector256<int>.Count)
+                    {
+                        Vector256<int> vec = Avx.LoadVector256(ptr + i);
+                        Avx.Store(ptr + i, Avx2.Xor(vec, ones));
+                    }
+                }
+            }
+            else if (Sse2.IsSupported)
+            {
+                Vector128<int> ones = Vector128.Create(-1);
+                fixed (int* ptr = thisArray)
+                {
+                    for (; i < count - (Vector128<int>.Count - 1); i += Vector128<int>.Count)
+                    {
+                        Vector128<int> vec = Sse2.LoadVector128(ptr + i);
+                        Sse2.Store(ptr + i, Sse2.Xor(vec, ones));
+                    }
+                }
+            }
+
+            for (; i < count; i++)
+                thisArray[i] = ~thisArray[i];
+
+        Done:
             _version++;
             return this;
         }