--- a/src/System.Collections/tests/BitArray/BitArray_CtorTests.cs
+++ b/src/System.Collections/tests/BitArray/BitArray_CtorTests.cs
@@ -76,12 +76,22 @@ namespace System.Collections.Tests
 
         public static IEnumerable<object[]> Ctor_BoolArray_TestData()
         {
+            Random rnd = new Random(0);
+
             yield return new object[] { new bool[0] };
             foreach (int size in new[] { 1, BitsPerByte, BitsPerByte * 2, BitsPerInt32, BitsPerInt32 * 2 })
             {
                 yield return new object[] { Enumerable.Repeat(true, size).ToArray() };
                 yield return new object[] { Enumerable.Repeat(false, size).ToArray() };
                 yield return new object[] { Enumerable.Range(0, size).Select(x => x % 2 == 0).ToArray() };
+
+                bool[] random = new bool[size];
+                for (int i = 0; i < random.Length; i++)
+                {
+                    random[i] = rnd.Next(0, 2) == 0;
+                }
+
+                yield return new object[] { random };
             }
         }