… src/System.Collections/tests/BitArray/BitArray_GetSetTests.cs
… src/System.Collections/tests/BitArray/BitArray_OperatorsTests.cs