@adamsitnik 2019-11-06
@tannergooding @BruceForstall could you please take a look? I would love to merge it before repo consolidation